// v2/api-validator/src/client/generated/models/BadRequestErrorType.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export enum BadRequestErrorType {
    SCHEMA_ERROR = 'schema-error',
    SCHEMA_PROPERTY_ERROR = 'schema-property-error',
    BAD_REQUEST = 'bad-request',
    UNKNOWN_ASSET = 'unknown-asset',
    UNSUPPORTED_CONVERSION = 'unsupported-conversion',
    QUOTE_NOT_READY = 'quote-not-ready',
    DEPOSIT_ADDRESS_DISABLED = 'deposit-address-disabled',
    ORDER_NOT_TRADING = 'order-not-trading',
    IDEMPOTENCY_KEY_REUSE = 'idempotency-key-reuse',
    DEPOSIT_ADDRESS_CREATION_NOT_ALLOWED = 'deposit-address-creation-not-allowed',
    UNSUPPORTED_TRANSFER_METHOD = 'unsupported-transfer-method',
    TRANSFER_DESTINATION_NOT_ALLOWED = 'transfer-destination-not-allowed',
    UNSUPPORTED_RAMP_METHOD = 'unsupported-ramp-method',
    UNSUPPORTED_SOURCE_ASSET = 'unsupported-source-asset',
    UNSUPPORTED_DESTINATION_ASSET = 'unsupported-destination-asset',
    AMOUNT_BELOW_MINIMUM = 'amount-below-minimum',
    PII_MISSING = 'pii-missing',
    UNSUPPORTED_EXTERNAL_SOURCE = 'unsupported-external-source',
    UNSUPPORTED_REGION = 'unsupported-region',
    DESTINATION_NOT_WHITELISTED = 'destination-not-whitelisted',
}